# back/config/swaggerDefinition.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "WiFiEvent_API",
    "description" : ""
  },
  "servers" : [ {
    "description" : "CRUD API",
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
...
    "/user/email" : {
      "post" : {
        "summary" : "send email authentication code",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "email" : {
                    "type" : "string",
                    "description" : "The user email."
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Send email success"
          },
          "500" : {
            "description" : "server error"
          }
        }
      }
    },
    "/user/auth" : {
      "post" : {
        "summary" : "check email authentication code",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "number" : {
                    "type" : "integer",
                    "description" : "Authentication code"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Email authentication success"
          },
          "401" : {
            "description" : "Wrong number"
          }
        }
      }
    },
...
  }
}